% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM-copy-coerce.R
\docType{methods}
\name{as.matrix,FBM-method}
\alias{as.matrix,FBM-method}
\title{Convert to base R matrix}
\usage{
\S4method{as.matrix}{FBM}(x)
}
\arguments{
\item{x}{A \link[=FBM-class]{FBM} object.}
}
\description{
Extract values from a Filebacked Big Matrix and convert to a base R matrix.
}
