% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInter.R
\name{big_addInter}
\alias{big_addInter}
\title{Augment a dataset with interactions}
\usage{
big_addInter(X., arr.ind, fun.createBM = BM())
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{arr.ind}{Matrix of two columns \code{row} and \code{col} specifying pairs
that need to be multiplied to create new columns.}

\item{fun.createBM}{A function that creates a \code{big.matrix}. It should have
exactly 3 parameters \code{nrow}, \code{ncol} and \code{type}. See \link{FBM}.}
}
\value{
The newly created \code{big.matrix} (or its descriptor).
}
\description{
Augment the number of column of a \code{big.matrox} by adding new columns
which are interactions between pairs. It may be good idea to use
\link{big_standardize} to have scaled columns as input.
}
\examples{
tmp <- tmpFBM(descriptor = FALSE)(10, 5, type = "double")
tmp[] <- rnorm(length(tmp))
big_standardize(tmp)
apply(tmp[,], 2, function(x) c(mean(x), sd(x)))

test <- big_addInter(tmp, arr.ind = cbind(1:4, 5))
}
