% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach-descriptor.R
\name{big_address}
\alias{big_address}
\title{Address of the described "big.matrix"}
\usage{
big_address(X.desc)
}
\arguments{
\item{X.desc}{A \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.
You shouldn't have missing values in your data.}
}
\value{
The slot \code{address} of the described \code{big.matrix}.
}
\description{
Address of the described "big.matrix"
}
\examples{
tmpfile <- tempfile()
test <- big.matrix(10, 10, backingfile = basename(tmpfile),
                   backingpath = dirname(tmpfile),
                   descriptorfile = paste0(basename(tmpfile), ".desc"))
X.desc <- describe(test)
address <- big_address(X.desc)
all.equal(address, test@address)
}
