% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{big_apply}
\alias{big_apply}
\title{Split-Apply-Combine}
\usage{
big_apply(X, FUN, .combine, block.size = 1000, ncores = 1)
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{FUN}{The function to be applied to each column block matrix.}

\item{.combine}{function that is used to process the tasks results as
    they generated.  This can be specified as either a function or
    a non-empty character string naming the function.
    Specifying 'c' is useful for concatenating the results into
    a vector, for example.  The values 'cbind' and 'rbind' can combine
    vectors into a matrix.  The values '+' and '*' can be used to
    process numeric data.
    By default, the results are returned in a list.}

\item{block.size}{Maximum number of columns read at once.
Default is \code{1000}.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}
}
\value{
The result of \link{foreach}.
}
\description{
A Split-Apply-Combine strategy to apply common R functions to a \code{big.matrix}.
}
\details{
This function split a \code{big.matrix} in column blocks, then apply a given
function to each block matrix and finally combine the results.
}
