% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach-descriptor.R
\name{big_attach}
\alias{big_attach}
\alias{big_attachExtdata}
\title{Attach a Filebacked Big Matrix}
\usage{
big_attach(rdsfile)

big_attachExtdata()
}
\arguments{
\item{rdsfile}{Path to a ".rds" file.}
}
\value{
The \link{FBM} object stored in the rdsfile.
}
\description{
Attach a Filebacked Big Matrix
}
\examples{
# tmpFBM
tmpfile <- tempfile()
X.desc <- FBM(backingroot = basename(tmpfile),
              backingpath = dirname(tmpfile))(10, 10)

descriptorfile <- paste0(tmpfile, ".desc")
X.desc2 <- big_attach(descriptorfile)

all.equal(X.desc, X.desc2)
}
\keyword{internal}
