% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM-copy-coerce.R
\name{big_copy}
\alias{big_copy}
\title{Title}
\usage{
big_copy(X, ind.row = rows_along(X), ind.col = cols_along(X),
  type = X$description$type, ..., block.size = block_size(length(ind.row)),
  ncores = 1)
}
\arguments{
\item{ncores}{}
}
\description{
Title
}
