% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{big_counts}
\alias{big_counts}
\title{Counts}
\usage{
big_counts(X.code, ind.row = rows_along(X.code),
  ind.col = cols_along(X.code), byrow = FALSE)
}
\arguments{
\item{X.code}{Either a \link[=BM.code-class]{BM.code} or
a \link[=BM.code.descriptor-class]{BM.code.descriptor}.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. \strong{Don't use negative indices.}}
}
\value{
A matrix of counts of K x m (or n) elements, where
\itemize{
\item K is the number of unique elements of the \code{BM.code},
\item n is its number of rows,
\item m is its number of columns.
}

\strong{Beware that K is up to 256. So, if you apply this on a \code{big.matrix} of
one million columns, you will create a matrix of nearly 1Gb!}.
}
\description{
Counts by columns (or rows) the number of each unique element of a \code{BM.code}.
}
\examples{
set.seed(11)

a <- matrix(as.raw(0), 14, 11)
a[] <- sample(as.raw(0:3), size = length(a), replace = TRUE)
X <- as.big.matrix(a)
X[,]

code <- rep(NA_real_, 256)
code[1:4] <- c(2, 5, 9, NA)
X2 <- as.BM.code(X, code)
X2[,]

# by columns
big_counts(X2)
apply(X2[,], 2, table, exclude = NULL)

# by rows
big_counts(X2, byrow = TRUE)
apply(X2[,], 1, table, exclude = NULL)

}
