% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mult.R
\name{big_cprodVec}
\alias{big_cprodVec}
\title{Cross-product between a "big.matrix" and a vector}
\usage{
big_cprodVec(X, y, ind.row = seq(nrow(X)), ind.col = seq(ncol(X)))
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used.}
}
\value{
\eqn{X^T /cdot y}.
}
\description{
Cross-product between a "big.matrix" and a vector
}
\examples{
N <- 100
M <- 20
X <- big.matrix(N, M)
X[] <- rnorm(length(X))
y <- rnorm(M)

test <- big_cprodVec(X, y) # vector
true <- crossprod(X[,], y) # one-column matrix
print(all.equal(test, as.numeric(true)))

# subsetting
ind.row <- sample(N, N/2)
ind.col <- sample(M, M/2)

# test2 <- big_prodVec(X, y, ind.row, ind.col)
# returns an error. You need to use the subset of y:
test2 <- big_prodVec(X, y[ind.col], ind.row, ind.col)
true2 <- X[ind.row, ind.col] \%*\% y[ind.col]
print(all.equal(test2, as.numeric(true2)))
}
