% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprodSelf2.R
\name{big_crossprodSelf2}
\alias{big_crossprodSelf2}
\title{Crossprod}
\usage{
big_crossprodSelf2(X, fun.scaling, ind.row = rows_along(X),
  ind.col = cols_along(X), block.size = block_size(nrow(X)))
}
\arguments{
\item{X}{A \link[=FBM-class]{FBM}.}

\item{fun.scaling}{A function that returns a named list of
\code{mean} and \code{sd} for every column, to scale each of their elements
such as followed: \deqn{\frac{X_{i,j} - mean_j}{sd_j}.}}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. \strong{Don't use negative indices.}}

\item{block.size}{Maximum number of columns read at once.
Default uses \link{block_size}.}
}
\value{
A matrix, \eqn{X.row^T X.row}, with the following two attributes:
\itemize{
\item a numeric vector \code{mean} of column scaling,
\item a numeric vector \code{sd} of column scaling.
}
}
\description{
Compute \eqn{X.row^T X.row} for a Filebacked Big Matrix \code{X}
after applying a particular scaling to it.
}
\details{
This algorithm is not really memory efficient. I'm planning on fixing this,
if it is really needed.
}
\section{Matrix parallelization}{

Large matrix computations (crossprods) are made block-wise and won't
be parallelized in order to not have to reduce the size of these blocks.
Instead, you may use \href{https://mran.microsoft.com/open/}{Microsoft R Open}
in order to accelerate these block matrix computations.
}

\examples{
X <- FBM(13, 17, init = rnorm(221))

# Comparing with tcrossprod
big_noscale <- big_scale(center = FALSE)
K <- big_crossprodSelf(X, fun.scaling = big_noscale)
dim(K)

true <- crossprod(X[])
all.equal(K, true, check.attributes = FALSE)

# Using only half of the data
n <- nrow(X)
ind <- sort(sample(n, n/2))
K2 <- big_crossprodSelf(X, fun.scaling = big_noscale, ind.row = ind)
dim(K2)

true2 <- crossprod(X[ind, ])
all.equal(K2, true2, check.attributes = FALSE)


}
\seealso{
\link{crossprod}
}
