% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{big_funScaling}
\alias{big_funScaling}
\alias{big_center}
\alias{big_scale}
\alias{big_noscale}
\title{Some scaling functions.}
\usage{
big_center(X, ind.train = seq(nrow(X)))

big_scale(X, ind.train = seq(nrow(X)))

big_noscale(X, ind.train = seq(nrow(X)))
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}
}
\value{
A named list of two vectors "mean" and "sd" which are as long as
the number of columns of \code{X}. Equivalence with \link{scale}:
\itemize{
\item \code{big_noscale}: \code{center = FALSE} & \code{scale = FALSE},
\item \code{big_center}:  \code{center = TRUE}  & \code{scale = FALSE},
\item \code{big_scale}:   \code{center = TRUE}  & \code{scale = TRUE}.
}
}
\description{
Some scaling functions for a "big.matrix" to be used as
the \code{fun.scaling} parameter of some functions of this package.
}
\examples{
print(big_center)
print(big_scale)
print(big_noscale)
}

