% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-parallelize.R
\name{big_parallelize}
\alias{big_parallelize}
\title{Split-parApply-Combine}
\usage{
big_parallelize(X.desc, p.MARGIN, p.FUN, p.combine, ncores,
  ind.row = seq_len(nrow(X.desc)), ind.col = seq_len(ncol(X.desc)), ...)
}
\arguments{
\item{X.desc}{A \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.
You shouldn't have missing values in your data.}

\item{p.FUN}{The function to be applied. \code{X.desc} must be its first argument
and it should have two other arguments, \code{ind.row} and \code{ind.col}.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used.}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used.}

\item{...}{Extra arguments to be passed to \code{FUN}.}
}
\value{
The result of \link{foreach}.
}
\description{
A Split-Apply-Combine strategy to parallelize the evaluation of a function.
}
\details{
This function splits indices in parts, then apply a given function to each
part and finally combine the results.
}
\examples{
X.desc <- big_attachExtdata()

test <- big_colstats(X.desc)
test2 <- big_parallelize(X.desc, 2, big_colstats,
                         p.combine = 'rbind', ncores = 2)

all.equal(test2, test)
}
