% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-parallelize.R
\name{big_parApply}
\alias{big_parApply}
\title{Split-parApply-Combine}
\usage{
big_parApply(X.desc, a.FUN, a.combine, ncores, block.size = 1000,
  ind = cols_along(X.desc), ...)
}
\arguments{
\item{X.desc}{A \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.
You shouldn't have missing values in your data.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}

\item{block.size}{Maximum number of columns read at once.
Default is \code{1000}. This parameter controls the trade-off between
memory usage and speed. Basically, the more you can load at once,
the quicker will be the execution time, at the expense of memory usage.}

\item{...}{Extra arguments to be passed to \code{FUN}.}

\item{p.FUN}{The function to be applied. \code{X.desc} must be its first argument
and it should have two other arguments, \code{ind.row} and \code{ind.col}.}
}
\value{
The result of \link{foreach}.
}
\description{
A Split-Apply-Combine strategy to parallelize the evaluation of
\code{big_apply} on a function.
}
\details{
This function splits indices in parts for parallelization, then split again
them on each core, apply a given function to each part and finally combine
the results (on each cluster and then from each cluster).
}
\examples{
X.desc <- big_attachExtdata()

test <- big_colstats(X.desc)
test2 <- big_parallelize(X.desc, 2, big_colstats,
                         p.combine = 'rbind', ncores = 2)

all.equal(test2, test)
}
