% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_Lanczos.R
\name{big_randomSVD2}
\alias{big_randomSVD2}
\title{A randomized algorithm for SVD.}
\usage{
big_randomSVD2(X, fun.scaling, ind.train = seq(nrow(X)), block.size = 1000,
  K = 10, I = 10, use.Eigen = TRUE)
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{fun.scaling}{A function that returns a named list of
\strong{\code{mean}} and \strong{\code{sd}} for every column, to scale each of their elements
such as followed: \deqn{\frac{X_{i,j} - mean_j}{sd_j}}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}

\item{block.size}{Maximum number of columns read at once.
Default is \code{1000}.}

\item{K}{Number of PCs to compute. This algorithm shouldn't
be used to compute a lot of PCs. Default is \code{10}.}

\item{I}{The number of iterations of the algorithm. Default is \code{10}.}

\item{use.Eigen}{Should the \code{Eigen} library be used
for matrix computations? Default tries to detect MRO. See details.}

\item{backingpath}{If \code{X} is filebacked and parallelism is used,
the path where are stored the files that are backing \code{X}.}
}
\description{
A randomized algorithm for SVD (or PCA) of a "big.matrix".
}
\references{
Rokhlin, V., Szlam, A., & Tygert, M. (2010).
A Randomized Algorithm for Principal Component Analysis.
SIAM Journal on Matrix Analysis and Applications, 31(3), 1100–1124.
doi:10.1137/080736417
}
\seealso{
\link{big_funScaling} \link{prcomp} \link{svd}
}
