% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{big_scale}
\alias{big_scale}
\title{Some scaling functions}
\usage{
big_scale(center = TRUE, scale = TRUE)
}
\arguments{
\item{center}{A logical value: whether to return means or 0s.}

\item{scale}{A logical value: whether to return sds or 1s. \strong{You can't
use scale without using center.}}
}
\value{
A new function that returns a named list of two vectors "mean" and "sd"
which are as long as the number of columns of \code{X}.
}
\description{
Some scaling functions for a "big.matrix" to be used as
the \code{fun.scaling} parameter of some functions of this package.
}
\seealso{
\link{scale}
}

