% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglassoLog.R
\name{big_spRegLogCV2}
\alias{big_spRegLogCV2}
\title{Title}
\usage{
big_spRegLogCV2(X, y01.train, ind.train = seq(nrow(X)), covar.train = NULL,
  K = 10, ncores = 1, ...)
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{y01.train}{Vector of responses, corresponding to \code{ind.train}.
Must be 0s and 1s.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all rows are used.}

\item{covar.train}{Matrix of covariables to be added in each model to correct
for confounders (e.g. the scores of PCA), corresponding to \code{ind.train}.
Default is \code{NULL} and corresponds to only adding an Intercept to each model.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}

\item{...}{}
}
\description{
Title
}
