% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colstats.R
\name{big_standardize}
\alias{big_standardize}
\title{Standardize a double "big.matrix"}
\usage{
big_standardize(X., thr.sd = 1e-04)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{thr.sd}{Threshold on standard deviation under which to ignore a column.}
}
\value{
A vector of column indices that have a low standard deviation.
}
\description{
Standardize a double "big.matrix"
}
\examples{
tmp <- tmpFBM(descriptor = FALSE)(10, 5, type = "double")
tmp[] <- rnorm(length(tmp))
apply(tmp[,], 2, function(x) c(mean(x), sd(x)))

big_standardize(tmp)
apply(tmp[,], 2, function(x) c(mean(x), sd(x)))
}
