% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrossprodSelf.R
\name{big_tcrossprodSelf}
\alias{big_tcrossprodSelf}
\title{Tcrossprod}
\usage{
big_tcrossprodSelf(X., fun.scaling, ind.row = rows_along(X.),
  block.size = 1000)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{fun.scaling}{A function that returns a named list of
\strong{\code{mean}} and \strong{\code{sd}} for every column, to scale each of their elements
such as followed: \deqn{\frac{X_{i,j} - mean_j}{sd_j}}.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. Don't use negative indices.}

\item{block.size}{Maximum number of columns read at once.
Default is \code{1000}. This parameter controls the trade-off between
memory usage and speed. Basically, the more you can load at once,
the quicker will be the execution time, at the expense of memory usage.}
}
\value{
A list of
\itemize{
\item \eqn{K = X.row X.row^T},
\item a numeric vector \code{mean} of column scaling,
\item a numeric vector \code{sd} of column scaling.
}
}
\description{
Compute \eqn{X.row X.row^T} for a \code{big.matrix} \code{X}
after applying a particular scaling to it.
}
\examples{
X.desc <- big_attachExtdata()

# Comparing with tcrossprod
big_noscale <- big_scale(center = FALSE)
test <- big_tcrossprodSelf(X.desc, fun.scaling = big_noscale)
dim(test$K)

true <- tcrossprod(attach.BM(X.desc)[,])
all.equal(test$K, true)

# Using only half of the data
n <- nrow(X.desc)
ind <- sort(sample(n, n/2))
test2 <- big_tcrossprodSelf(X.desc, fun.scaling = big_noscale,
                            ind.row = ind)
dim(test2$K)

true2 <- tcrossprod(attach.BM(X.desc)[ind, ])
all.equal(test2$K, true2)
}
\seealso{
\link{tcrossprod}
}
