% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{big_transpose}
\alias{big_transpose}
\title{Transposition}
\usage{
big_transpose(X., descriptor = TRUE, ...)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{descriptor}{Return the descriptor of the transposed \code{big.matrix} or
directly the \code{big.matrix} object? Default is \code{TRUE} (the descriptor).}

\item{...}{Arguments passed on to \code{bigmemory::big.matrix}
\describe{
  \item{separated}{use separated column organization of the data; 
see details.}
  \item{backingfile}{the root name for the file(s) for the cache of \code{x}.}
  \item{backingpath}{the path to the directory containing the file 
backing cache.}
  \item{descriptorfile}{the name of the file to hold the backingfile 
description, for subsequent use with \code{\link{attach.big.matrix}}; 
if \code{NULL}, the \code{backingfile} is used as the root part of the 
descriptor file name.  The descriptor file is placed in the same directory 
as the backing files.}
  \item{binarydescriptor}{the flag to specify if the binary RDS format 
should be used for the backingfile description, for subsequent use with 
\code{\link{attach.big.matrix}}; if \code{NULL} of \code{FALSE}, the 
\code{dput()} file format is used.}
  \item{shared}{\code{TRUE} by default, and always \code{TRUE} if the 
\code{big.matrix} is file-backed.  For a non-filebacked \code{big.matrix}, 
\code{shared=FALSE} uses non-shared memory, which can be more stable for 
large (say, >50\% of RAM) objects.  Shared memory allocation can sometimes 
fail in such cases due to exhausted shared-memory resources in the system.}
}}
}
\value{
The new transposed \code{big.matrix} (or its descriptor). Its dimension,
type and dimnames are automatically determined from the input \code{big.matrix}.
}
\description{
This function implements a simple cache-oblivious
algorithm for the transposition of a "big.matrix".
}
\examples{
X.desc <- big_attachExtdata()

tmpfile <- tempfile()
Xt.desc <- big_transpose(X.desc, descriptor = TRUE,
                         backingfile = basename(tmpfile),
                         backingpath = dirname(tmpfile),
                         descriptorfile = paste0(basename(tmpfile), ".desc"))

identical(t(attach.BM(X.desc)[,]), attach.BM(Xt.desc)[,])
}
