% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{big_yaware_scale}
\alias{big_yaware_scale}
\title{Y-aware scaling}
\usage{
big_yaware_scale(y)
}
\arguments{
\item{y}{Vector of responses.}
}
\value{
A new \strong{function} that returns a named list of
two vectors \strong{\code{mean}} and \strong{\code{sd}}
which are as long as the number of columns of \strong{\code{X}}.
}
\description{
Y-aware scaling
}
\details{
"y-aware" scaling center columns and multiply them
by betas of univariate linear regression.
So, \strong{\code{mean}} are column means and \strong{\code{sd}} are the inverse of slopes.
See \href{https://goo.gl/8G8WMa}{this blog post} for details.
}
\examples{
# Simulating some data
X <- big.matrix(41, 17)
X[] <- rnorm(length(X))
y <- rnorm(nrow(X), X[, 9], abs(X[, 9]))

X.svd <- bigstatsr::big_SVD(X, fun.scaling = big_yaware_scale(y))
plot(X.svd$v[, 1], type = "h")
}

