% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigXtX.R
\name{BigXtX}
\alias{BigXtX}
\title{Crossprod for a big.matrix.}
\usage{
BigXtX(X, block.size, ind.train = seq(nrow(X)), fun.scaling,
  use.Eigen = TRUE, progress = TRUE)
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{block.size}{Maximum number of columns read at once.
Default is \code{1000}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}

\item{use.Eigen}{Use the \code{Eigen} library to compute
\eqn{X^T X}? \code{TRUE} is the default.}

\item{progress}{}
}
\value{
\eqn{X.train^T X.train}
}
\description{
Compute \eqn{X^T X} for a "big.matrix" X
after applying a particular scaling to it.
}
\seealso{
\link{crossprod}
}

