% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-new-methods.R
\docType{class}
\name{BM.code-class}
\alias{BM.code-class}
\alias{BM.code}
\alias{BM.code.descriptor-class}
\alias{as.BM.code}
\alias{as.BM.code,big.matrix,numeric-method}
\alias{as.BM.code,big.matrix.descriptor,numeric-method}
\title{S4 class BM.code}
\usage{
as.BM.code(x, code)

\S4method{as.BM.code}{big.matrix,numeric}(x, code)

\S4method{as.BM.code}{big.matrix.descriptor,numeric}(x, code)
}
\arguments{
\item{x}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{code}{A numeric vector of length 256. You should contruct it with
\code{rep(NA_real_, 256)} and then replace the values
which are of interest for you.}
}
\description{
A class for storing and accessing up to 256 values using a RAW "big.matrix".
}
\details{
Compared to a \link[=big.matrix-class]{big.matrix}, it adds a slot \code{code} which
is used as a lookup table of size 256.
}
\examples{
X <- big.matrix(10, 10, type = "raw")
X[] <- sample(as.raw(0:3), size = length(X), replace = TRUE)
X[,]

code <- rep(NA_real_, 256)
code[1:3] <- c(1, 3, 5)

X.code <- as.BM.code(X, code)
X.code[,]

}
