% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_X}
\alias{check_X}
\title{Checking of objects}
\usage{
check_X(X, y = NULL, y.type = "null")
}
\arguments{
\item{X}{A matrix-like object you want to check as being a
\link[=big.matrix-class]{big.matrix}.}

\item{y}{A vector of responses you want to check compatibility with
what you are doing (either classification or regression).}

\item{y.type}{A string:
\itemize{
\item "reg" is you want to check the compatibility with regression,
\item "class" is you want to check the compatibility with classification.
}}
}
\value{
Throwing an error if something is wrong.
}
\description{
Check if X is a \code{big.matrix} and
possibly if y is compatible with classification or regression.
}
\examples{
\dontrun{

  X <- big.matrix(4, 4)
  check_X(X) # OK
  mat <- as.matrix(X)
  check_X(mat) # NOK

  y <- rnorm(4)
  check_X(X, y, y.type = "reg") # OK
  check_X(X, y, y.type = "class") # NOK
}
}

