% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colstats.R
\name{colstats}
\alias{colmeans}
\alias{colmeans_sds}
\alias{colsds}
\alias{colstats}
\alias{colsums}
\alias{colvars}
\title{Standard statistics for \strong{columns} of a big.matrix}
\usage{
colsums(X, ind.train = seq_len(nrow(X)))

colmeans(X, ind.train = seq_len(nrow(X)))

colvars(X, ind.train = seq_len(nrow(X)))

colsds(X, ind.train = seq_len(nrow(X)))

colmeans_sds(X, ind.train = seq_len(nrow(X)))
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part.
If not specified, all data are used.}
}
\value{
Numeric vectors with the column statistics.
}
\description{
Standard statistics for \strong{columns} of a big.matrix.
For now, the \code{sum}, \code{mean}, \code{sd} and \code{var} are implemented.
}
\examples{
# Simulating some data
X <- big.matrix(10, 10, shared = FALSE)
X[] <- rnorm(100)

# Check the results
print(test <- colsums(X))
print(test - colSums(X[,]))

# Only with the first 5 rows
ind <- 1:5
print(test <- colsums(X, ind))
print(test - colSums(X[ind, ]))

# NA in, NA out
X[2, 5] <- NA
print(colsums(X))
}

