% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-new-methods.R
\docType{methods}
\name{typeof,big.matrix.descriptor-method}
\alias{typeof,big.matrix.descriptor-method}
\alias{describe,big.matrix.descriptor-method}
\alias{describe,BM.code-method}
\alias{attach.BM}
\alias{attach.BM,big.matrix-method}
\alias{attach.BM,big.matrix.descriptor-method}
\alias{attach.BM,BM.code.descriptor-method}
\title{Extend "bigmemory" methods}
\usage{
\S4method{typeof}{big.matrix.descriptor}(x)

\S4method{describe}{big.matrix.descriptor}(x)

\S4method{describe}{BM.code}(x)

attach.BM(x)

\S4method{attach.BM}{big.matrix}(x)

\S4method{attach.BM}{big.matrix.descriptor}(x)

\S4method{attach.BM}{BM.code.descriptor}(x)
}
\arguments{
\item{x}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}
}
\description{
Extend some methods for completeness, so that \code{big.matrix} objects
or their descriptor can be used interchangeably.
}
