% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglasso.R
\name{COPY_biglasso_part}
\alias{COPY_biglasso_part}
\title{Train one model}
\usage{
COPY_biglasso_part(
  X,
  y.train,
  ind.train,
  ind.col,
  covar.train,
  family,
  lambda,
  center,
  scale,
  resid,
  alpha,
  eps,
  max.iter,
  dfmax,
  ind.val,
  covar.val,
  y.val,
  n.abort,
  nlam.min,
  base.train,
  base.val,
  pf,
  weights.train,
  weights.val
)
}
\value{
A named list with following variables:
\item{intercept}{A vector of intercepts, corresponding to each lambda.}
\item{beta}{The vector of coefficients that minimized the loss on the
validation set.}
\item{iter}{A vector of length \code{nlambda} containing the number of
iterations until convergence at each value of \code{lambda}.}
\item{lambda}{The sequence of regularization parameter values in the path.}
\item{alpha}{Input parameter.}
\item{loss}{A vector containing either the residual sum of squares
(for linear models) or negative log-likelihood (for logistic models)
of the fitted model at each value of \code{lambda}.}
\item{loss.val}{A vector containing the loss for the corresponding
validation set.}
\item{message}{Reason the fitting has stopped.}
\item{nb_active}{The number of active (non-zero) variables along the
regularization path.}
\item{nb_candidate}{The number of candidate variables (used in the gradient
descent) along the regularization path.}
\item{ind.train}{Indices of training set.}
}
\description{
Train one model
}
\keyword{internal}
