% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{covar_from_df}
\alias{covar_from_df}
\title{Numeric matrix from data frame}
\usage{
covar_from_df(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A numeric matrix.
}
\description{
Transform a data frame to a numeric matrix by one-hot encoding factors.
The last factor value is always omitted to prevent having a singular matrix
when adding a column of 1s (intercept) in models.
}
\examples{
mat <- covar_from_df(iris)
head(mat)
}
