% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBM.R
\name{BM}
\alias{BM}
\alias{FBM}
\alias{tmpFBM}
\title{Functions for creating a "big.matrix"}
\usage{
BM(..., descriptor = TRUE)

FBM(backingroot, backingpath = "backingfiles", init = NULL,
  descriptor = TRUE)

tmpFBM(init = NULL, descriptor = TRUE)
}
\arguments{
\item{...}{Arguments passed on to \code{bigmemory::big.matrix}
\describe{
  \item{init}{a scalar value for initializing the matrix (\code{NULL} by 
default to avoid unnecessary time spent doing the initializing).}
  \item{dimnames}{a list of the row and column names; use with caution 
for large objects.}
  \item{separated}{use separated column organization of the data; 
see details.}
  \item{backingfile}{the root name for the file(s) for the cache of \code{x}.}
  \item{backingpath}{the path to the directory containing the file 
backing cache.}
  \item{descriptorfile}{the name of the file to hold the backingfile 
description, for subsequent use with \code{\link{attach.big.matrix}}; 
if \code{NULL}, the \code{backingfile} is used as the root part of the 
descriptor file name.  The descriptor file is placed in the same directory 
as the backing files.}
  \item{binarydescriptor}{the flag to specify if the binary RDS format 
should be used for the backingfile description, for subsequent use with 
\code{\link{attach.big.matrix}}; if \code{NULL} of \code{FALSE}, the 
\code{dput()} file format is used.}
  \item{shared}{\code{TRUE} by default, and always \code{TRUE} if the 
\code{big.matrix} is file-backed.  For a non-filebacked \code{big.matrix}, 
\code{shared=FALSE} uses non-shared memory, which can be more stable for 
large (say, >50\% of RAM) objects.  Shared memory allocation can sometimes 
fail in such cases due to exhausted shared-memory resources in the system.}
}}

\item{descriptor}{Return the descriptor of the resulting \code{big.matrix} or
directly the \code{big.matrix} object? Default is \code{TRUE} (the descriptor).}

\item{backingroot}{The root of the backing file to be created.
This should not be a full path (use \code{backingpath})
nor have any extension (".bk" is used by default).}

\item{backingpath}{The directory of the backing file to be created.
Default is \code{"backingfiles"}. The directory is created if doesn't exist.}

\item{init}{A scalar value for initializing the matrix
(\code{NULL} by default to avoid unnecessary time spent doing the initializing).}
}
\value{
A \strong{function} of 3 parameters \code{nrow}, \code{ncol} and \code{type} (default is
\code{"double"}) that is used to create a \code{big.matrix}.
This is particularly useful as the \code{fun.createBM} parameter.
}
\description{
The 3 following functions return \strong{functions} as well:
\itemize{
\item \code{BM}: wrapper around \link{big.matrix}.
}

\itemize{
\item \code{FBM}: wrapper around \link{filebacked.big.matrix}. In particular:
\itemize{
\item backingpath is \code{"backingfiles"} by default,
\item extension \code{.bk} is used,
\item descriptor is never binary and has an extension ".desc",
\item the \code{big.matrix} has no dimname,
\item option \code{separated} is never used.
}
}

\itemize{
\item \code{tmpFBM}: wrapper around \code{FBM} to get a temporary filebacked \code{big.matrix}.
It uses \link{tempfile} to get the location of
the temporary filebacked \code{big.matrix}.
Temporary means not remaining if you restart your R session.
}
}
\examples{
# these functions return **functions**
(f <- BM(shared = FALSE, descriptor = FALSE))
# with 3 parameters
(X <- f(nrow = 10, ncol = 6, type = "integer"))
is.shared(X)

# a shortcut for a filebacked `big.matrix`
X2.desc <- FBM(backingroot = "test", init = 2)(10, 10)
str(X2.desc)
attach.BM(X2.desc)[,]
file.remove(paste0("backingfiles/test", c(".bk", ".desc")))

# a temporary filebacked `big.matrix`
test <- tmpFBM()(10, 10)
str(test)
}
