% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-new-methods.R
\docType{methods}
\name{nrow,big.matrix.descriptor-method}
\alias{nrow,big.matrix.descriptor-method}
\alias{ncol,big.matrix.descriptor-method}
\alias{dim,big.matrix.descriptor-method}
\alias{length,big.matrix.descriptor-method}
\title{Dimensions}
\usage{
\S4method{nrow}{big.matrix.descriptor}(x)

\S4method{ncol}{big.matrix.descriptor}(x)

\S4method{dim}{big.matrix.descriptor}(x)

\S4method{length}{big.matrix.descriptor}(x)
}
\arguments{
\item{x}{A \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}
}
\description{
Dimensions of the described \code{big.matrix} object. This can be used to know
the dimensions of the underlying \code{big.matrix} without having to attach it.
}
\examples{
X.desc <- big_attachExtdata()
dim(X.desc)
dim(attach.big.matrix(X.desc))

}
