% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-new-methods.R
\docType{methods}
\name{[,BM.code,missing,missing,missing-method}
\alias{[,BM.code,missing,missing,missing-method}
\alias{[,BM.code,missing,missing,logical-method}
\alias{[,BM.code,missing,numeric,missing-method}
\alias{[,BM.code,missing,numeric,logical-method}
\alias{[,BM.code,numeric,missing,missing-method}
\alias{[,BM.code,numeric,missing,logical-method}
\alias{[,BM.code,numeric,numeric,missing-method}
\alias{[,BM.code,numeric,numeric,logical-method}
\alias{[,BM.code,matrix,missing,missing-method}
\title{Extract}
\usage{
\S4method{[}{BM.code,missing,missing,missing}(x, i, j, drop)

\S4method{[}{BM.code,missing,missing,logical}(x, i, j, drop)

\S4method{[}{BM.code,missing,numeric,missing}(x, i, j, drop)

\S4method{[}{BM.code,missing,numeric,logical}(x, i, j, drop)

\S4method{[}{BM.code,numeric,missing,missing}(x, i, j, drop)

\S4method{[}{BM.code,numeric,missing,logical}(x, i, j, drop)

\S4method{[}{BM.code,numeric,numeric,missing}(x, i, j, drop)

\S4method{[}{BM.code,numeric,numeric,logical}(x, i, j, drop)

\S4method{[}{BM.code,matrix,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A \link{BM.code}.}

\item{i}{Indices specifying the rows.}

\item{j}{Indices specifying the columns.}

\item{drop}{Logical indication if reduce to minimum dimensions.}
}
\description{
Extract replace big.matrix elements
}
\keyword{internal}
