% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM-code256.R
\docType{methods}
\name{FBM.code256-methods}
\alias{FBM.code256-methods}
\alias{[,FBM.code256,ANY,ANY,ANY-method}
\title{Methods for the FBM.code256 class}
\usage{
\S4method{[}{FBM.code256,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \link[=FBM.code256-class]{FBM.code256}.}

\item{i}{A vector of indices (or nothing). You can use positive and negative
indices, logical indices (that are recycled) and also a matrix of indices
(but only positive ones).}

\item{j}{A vector of indices (or nothing). You can use positive and negative
indices, logical indices (that are recycled).}

\item{...}{Not used. Just to make \link{nargs} works.}

\item{drop}{Whether to delete the dimensions of a matrix which have
one dimension equals to 1.}
}
\description{
Methods for the FBM.code256 class

Accessor method for class \code{FBM.code256}.
}
