% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglasso.R
\name{get_beta}
\alias{get_beta}
\title{Combine sets of coefficients}
\usage{
get_beta(betas, method = c("geometric-median", "mean-wise", "median-wise"))
}
\arguments{
\item{betas}{Matrix of coefficient vectors to be combined.}

\item{method}{Method for combining vectors of coefficients. The default uses
the \href{https://en.wikipedia.org/wiki/Geometric_median}{geometric median}.}
}
\value{
A vector of resulting coefficients.
}
\description{
Combine sets of coefficients
}
