% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linRegPcadapt}
\alias{linRegPcadapt}
\title{T-scores used in pcadapt}
\usage{
linRegPcadapt(BM, U, rowInd)
}
\arguments{
\item{BM}{A \code{BM.code} object.}

\item{U}{Matrix of left singular vectors (from partial SVD).}

\item{rowInd}{Vector of row indices of the \code{big.matrix} that are used.}
}
\value{
A matrix of t-scores where rows correspond to each SNP and
columns correspond to each left singular vector.
}
\description{
Compute matrix of t-scores (SNPs x scores) used in pcadapt.
}
\details{
This function is used by package bigsnpr.
}
\references{
Keurcien Luu and Michael Blum (2017).
pcadapt: Fast Principal Component Analysis for Outlier Detection.
R package version 3.0.4. https://CRAN.R-project.org/package=pcadapt.
}
\keyword{internal}
