% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-parallelize.R
\name{nb_cores}
\alias{nb_cores}
\title{Recommended number of cores to use}
\usage{
nb_cores()
}
\value{
The recommended number of cores to use.
}
\description{
This is base on the following rule: use only physical cores and if you have
only physical cores, leave one core for the OS/UI.
}
\examples{
nb_cores()
}
