% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{nlines}
\alias{nlines}
\title{Number of lines}
\usage{
nlines(file)
}
\arguments{
\item{file}{Path of the file.}
}
\value{
The number of lines as one integer.
}
\description{
Get the number of lines of a file with system command \code{wc -l}.
}
\examples{
csv <- tempfile(fileext = "csv")
write.csv(iris, csv, quote = FALSE, row.names = FALSE)
tryCatch(csv, error = function(e) print(e))
}
