% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelRandomProjPCA.R
\name{ParallelRandomProjPCA}
\alias{ParallelRandomProjPCA}
\title{Title}
\usage{
ParallelRandomProjPCA(X, block.size, ncores, K = 10, I = 10,
  backingpath = NULL, tmp.lock.name = "file.lock", TIME = 0.01)
}
\description{
Title
}

