% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{pcor}
\alias{pcor}
\title{Partial correlation}
\usage{
pcor(x, y, z, alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{z}{A data frame, which can contain characters or factors.}

\item{alpha}{Type-I error for the confidence interval (CI).
Default is \code{0.05}, corresponding to a 95\% CI.}
}
\value{
The partial correlation, and the lower and upper bounds of its CI.
}
\description{
Partial correlation between x and y, after having adjusted both for z.
}
\examples{
pcor(iris[[1]], iris[[2]], iris[-(1:2)])

}
