% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.big_sp_list}
\alias{plot.big_sp_list}
\title{Plot method}
\usage{
\method{plot}{big_sp_list}(x, coeff = 1, ...)
}
\arguments{
\item{x}{An object of class \code{big_sp_list}.}

\item{coeff}{Relative size of text. Default is \code{1}.}

\item{...}{Not used.}
}
\value{
A \code{ggplot2} object. You can plot it using the \code{print} method.
You can modify it as you wish by adding layers. You might want to read
\href{http://r4ds.had.co.nz/data-visualisation.html}{this chapter}
to get more familiar with the package \strong{ggplot2}.
}
\description{
Plot method for class \code{big_sp_list}.
}
