% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.big_SVD}
\alias{plot.big_SVD}
\title{Plot method}
\usage{
\method{plot}{big_SVD}(x, type = c("screeplot", "scores", "loadings"),
  nval = length(x$d), scores = c(1, 2), loading = 1, coeff = 1, ...)
}
\arguments{
\item{x}{An object of class \code{big_SVD}.}

\item{type}{Either
\itemize{
\item "screeplot": plot of decresing singular values (the default).
\item "scores": plot of the scores associated with 2 Principal Components.
\item "loadings": plot of loadings associated with 1 Principal Component.
}}

\item{nval}{Number of singular values to plot. Default plots all computed.}

\item{scores}{Vector of indices of the two PCs to plot. Default plots the
first 2 PCs.}

\item{loading}{Index of the unique PCA loading to plot. Default plots the
first vector of loadings.}

\item{coeff}{Relative size of text. Default is \code{1}.}

\item{...}{Not used.}
}
\value{
A \code{ggplot2} object. You can plot it using the \code{print} method.
You can modify it as you wish by adding layers. You might want to read
\href{http://r4ds.had.co.nz/data-visualisation.html}{this chapter}
to get more familiar with the package \strong{ggplot2}. \cr
Do you want dynamic plots? Go check \url{https://plot.ly/ggplot2}.
}
\description{
Plot method for class \code{big_SVD}.
}
\examples{
test <- big_attachExtdata()
svd <- big_SVD(test, big_scale(), k = 10)

# screeplots
plot(svd) # 3 PCs seems "significant"
plot(svd, coeff = 1.2) # larger font for papers

# scores plot
plot(svd, type = "scores") # first 2 PCs
plot(svd, type = "scores", scores = c(1, 3))
## add color (recall that this return a `ggplot2` object)
class(obj <- plot(svd, type = "scores"))
population <- rep(c("POP1", "POP2", "POP3"), c(143, 167, 207))
library(ggplot2)
print(obj2 <- obj + geom_point(aes(color = population)))
## change the place of the legend
print(obj3 <- obj2 + theme(legend.position = c(0.85, 0.17)))
## change the title and the labels of the axes
obj3 + ggtitle("Yet another title") + xlab("with an other 'x' label")

# loadings
plot(svd, type = "loadings", loading = 2)

# dynamic plots, require the package **plotly**
\dontrun{plotly::ggplotly(obj3)}
}
\seealso{
\link{big_SVD} and \link{big_randomSVD}.
}
