% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhtest.R
\name{plot.mhtest}
\alias{plot.mhtest}
\title{Plot method}
\usage{
\method{plot}{mhtest}(x, type = c("Manhattan", "Volcano"),
  main = paste(type, "plot"), cex = 0.5, pch = 19, ...)
}
\arguments{
\item{x}{An object of class \code{mhtest}.}

\item{type}{Either
\itemize{
\item "Manhattan": plot of the negative logarithm (in base 10) of p-values.
\item "Volcaco": plot of the negative logarithm of p-values against the
estimation of coefficients (e.g. betas in linear regression).
}}

\item{main}{The title of the plot. Default use the \code{type}.}

\item{cex}{Graphical parameter: relative size of points. Default is \code{0.5}.}

\item{pch}{Graphical parameter: type of points. Default uses "filled" points.}

\item{...}{Other arguments to be passed to \code{plot()}.}
}
\value{
\code{NULL}. Creates a plot.
}
\description{
Plot method for class \code{mhtest}.
}
\seealso{
\link{big_univLinReg} and \link{big_univLogReg}.
}
