% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-parallelize.R
\name{plus}
\alias{plus}
\title{Add}
\usage{
plus(...)
}
\arguments{
\item{...}{Multiple arguments to be added together}
}
\value{
\code{Reduce('+', list(...))}
}
\description{
Add multiple arguments
}
\examples{
plus(1:3, 4:6, 1:3)
}
