% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.big_CMSA}
\alias{predict.big_CMSA}
\title{Predict method}
\usage{
\method{predict}{big_CMSA}(object, X, ind.row = rows_along(X),
  covar.row = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{big_CMSA}.}

\item{X}{A \link[=FBM-class]{FBM}.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{covar.row}{Matrix of covariables to be added in each model to correct
for confounders (e.g. the scores of PCA), corresponding to \code{ind.row}.
Default is \code{NULL} and corresponds to only adding an intercept to each model.}

\item{...}{Not used.}
}
\value{
A vector of prediction scores, corresponding to \code{ind.row}.
}
\description{
Predict method for class \code{big_CMSA}.
}
\seealso{
\link{big_CMSA}
}
