% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.big_sp}
\alias{predict.big_sp}
\title{Predict method}
\usage{
\method{predict}{big_sp}(object, X, ind.row = seq_len(nrow(X)),
  covar.row = NULL, block.size = block.size, ncores2 = 1)
}
\arguments{
\item{object}{Object of class \code{big_spReg}.}

\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used.}

\item{covar.row}{Matrix of covariables to be added in each model to correct
for confounders (e.g. the scores of PCA), corresponding to \code{ind.row}.
Default is \code{NULL} and corresponds to only adding an intercept to each model.}

\item{block.size}{Maximum number of columns read at once.
Default is \code{1000}. This parameter controls the trade-off between
memory usage and speed. Basically, the more you can load at once,
the quicker will be the execution time, at the expense of memory usage.}

\item{ncores2}{Number of cores used. Default doesn't use parallelism.
For this function, use only half of the cores you have.}
}
\value{
A vector of scores, corresponding to \code{ind.row}.
}
\description{
Predict method which is applicable to
\link{big_spLinReg}, \link{big_spLogReg} and \link{big_spSVM}.
}
