% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mhtest}
\alias{predict.mhtest}
\title{Predict method}
\usage{
\method{predict}{mhtest}(object, scores = object$score, log10 = TRUE,
  ...)
}
\arguments{
\item{object}{An object of class \code{mhtest} from you get the probability
function with possibly pre-transformation of scores.}

\item{scores}{Raw scores (before transformation) that you want to transform
to p-values.}

\item{log10}{Are p-values returned on the \code{log10} scale? Default is \code{TRUE}.}

\item{...}{Not used.}
}
\value{
Vector of \strong{\code{log10(p-values)}} associated with \code{scores} and \code{object}.
}
\description{
Predict method for class \code{mhtest}.
}
\seealso{
\link{big_univLinReg} and \link{big_univLogReg}.
}
