% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PrimalBigPCA}
\alias{PrimalBigPCA}
\title{Principal Components of a "big.matrix".}
\usage{
PrimalBigPCA(X, block.size, k = NULL, ind.train = seq(nrow(X)),
  vec.center = rep(0, length(ind.train)), vec.scale = rep(1,
  length(ind.train)), thr.eigval = 0.001, progress = TRUE)
}
\description{
Get k or all Principal Components (PCs) of a \code{big.matrix}
which has more rows than columns.
}

