% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomProjPCA.R
\name{RandomProjPCA}
\alias{RandomProjPCA}
\title{Title}
\usage{
RandomProjPCA(X, fun.scaling, block.size = 1000, K = 10, I = 10,
  use.Eigen = TRUE)
}
\arguments{
\item{fun.scaling}{}
}
\description{
Title
}

