% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM.R, R/apply-parallelize.R, R/plot.R,
%   R/utils-assert.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{FBM}
\alias{as_FBM}
\alias{big_copy}
\alias{big_transpose}
\alias{big_increment}
\alias{FBM.code256}
\alias{add_code256}
\alias{big_counts}
\alias{big_attach}
\alias{big_attachExtdata}
\alias{big_read}
\alias{big_write}
\alias{rows_along}
\alias{cols_along}
\alias{nb_cores}
\alias{plus}
\alias{big_parallelize}
\alias{big_apply}
\alias{block_size}
\alias{plot_grid}
\alias{without_downcast_warning}
\alias{sub_bk}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bigparallelr}{\code{\link[bigparallelr]{cols_along}}, \code{\link[bigparallelr]{nb_cores}}, \code{\link[bigparallelr]{plus}}, \code{\link[bigparallelr]{rows_along}}}

  \item{cowplot}{\code{\link[cowplot]{plot_grid}}}

  \item{rmio}{\code{\link[rmio]{FBM}}, \code{\link[rmio]{FBM.code256}}, \code{\link[rmio]{add_code256}}, \code{\link[rmio]{as_FBM}}, \code{\link[rmio]{big_apply}}, \code{\link[rmio]{big_attach}}, \code{\link[rmio]{big_attachExtdata}}, \code{\link[rmio]{big_copy}}, \code{\link[rmio]{big_counts}}, \code{\link[rmio]{big_increment}}, \code{\link[rmio]{big_parallelize}}, \code{\link[rmio]{big_read}}, \code{\link[rmio]{big_transpose}}, \code{\link[rmio]{big_write}}, \code{\link[rmio]{block_size}}, \code{\link[rmio]{sub_bk}}, \code{\link[rmio]{without_downcast_warning}}}
}}

