% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rows_along}
\alias{rows_along}
\alias{cols_along}
\title{Sequence generation}
\usage{
rows_along(x)

cols_along(x)
}
\arguments{
\item{x}{Any object on which you can call \code{nrow} and \code{ncol}.}
}
\description{
Similar to \link{seq_along}, it creates sequences of size \code{nrow(x)} or \code{ncol(x)}.
}
\examples{
X <- big_attachExtdata()
dim(X)
str(rows_along(X))
str(cols_along(X))

}
\keyword{internal}
