% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subBM-class.R
\name{subBM-class}
\alias{subBM-class}
\alias{subBM}
\alias{nrow.subBM}
\alias{ncol.subBM}
\alias{dim.subBM}
\alias{length.subBM}
\title{S3 class subBM}
\usage{
nrow.subBM(x)

ncol.subBM(x)

\method{dim}{subBM}(x)

\method{length}{subBM}(x)
}
\description{
A class for representing a subset view of a \code{big.matrix}.
}
\details{
A named list with at 3 slots: \describe{
\item{desc}{The \code{big.matrix.descriptor} object.}
\item{rows}{Indices of rows to be used.}
\item{cols}{Indices of columns to be used.}
}
}
