% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.big_sp_list}
\alias{summary.big_sp_list}
\title{Summary method}
\usage{
\method{summary}{big_sp_list}(object, best.only = FALSE, sort = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{big_sp_list}.}

\item{best.only}{Whether to return only one row corresponding to the best
model? The best model is the one smallest \verb{$validation_loss}.}

\item{sort}{Whether to sort by \verb{$validation_loss}. Default is \code{FALSE}.}

\item{...}{Not used.}
}
\value{
A tibble with, for each \verb{$alpha}, a mean \verb{$validation_loss}, a mean
vector of coefficients \verb{$beta}, the corresponding number of non-zero
coefficients \verb{$nb_var}, and the reasons of method completion \verb{$message}.
}
\description{
Summary method for class \code{big_sp_list}.
}
