% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_RSpectra.R
\name{test_RSpectra}
\alias{test_RSpectra}
\title{Title}
\usage{
test_RSpectra(X, fun.scaling, ind.train = seq(nrow(X)), block.size = 1000,
  k = 10, use.Eigen = !detect_MRO())
}
\arguments{
\item{use.Eigen}{}
}
\description{
Title
}
