% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_bigstatsr}
\alias{theme_bigstatsr}
\title{Theme ggplot2}
\usage{
theme_bigstatsr(size.rel = 1)
}
\arguments{
\item{size.rel}{Relative size. Default is \code{1}.}
}
\description{
Theme ggplot2 used by this package.
}
\examples{
library(ggplot2)
(p <- ggplot(mapping = aes(x = 1:10, y = 1:10)) + geom_point())
p + theme_bw()
p + theme_bigstatsr()
}
