% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmpFBM}
\alias{tmpFBM}
\title{Descriptor of temporary filebacked "big.matrix"}
\usage{
tmpFBM(n, m, type = "double", init = NULL)
}
\arguments{
\item{n}{Number of rows.}

\item{m}{Number of columns.}

\item{type}{The type of the atomic element. Default is \code{double}.}

\item{init}{A scalar value for initializing the matrix. Default is \code{NULL},
which avoids unnecessary time spent doing the initializing.}
}
\value{
A descriptor of a \code{big.matrix}, filebacked in directory "/tmp/".
}
\description{
Descriptor of temporary filebacked "big.matrix"
}
\examples{
X.desc <- tmpFBM(10, 10)
str(X.desc)
}
\keyword{internal}
