% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBM.R
\name{tmpFBM.rm}
\alias{tmpFBM.rm}
\title{Remove temporary backing files}
\usage{
tmpFBM.rm(X.)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}
}
\value{
These functions return a logical vector indicating which
  operation succeeded for each of the files attempted.  Using a missing
  value for a file or path name will always be regarded as a failure.

  If \code{showWarnings = TRUE}, \code{file.create} will give a warning
  for an unexpected failure.
}
\description{
Remove temporary backing files
}
\examples{
list.files(path = tempdir())
X.desc <- tmpFBM()(10, 10)
list.files(path = tempdir())

tmpFBM.rm(X.desc)
list.files(path = tempdir())

}
