% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\title{Transposition of a "big.matrix".}
\usage{
transpose(X, ...)
}
\arguments{
\item{X}{A big.matrix. You may have missing values in your data.}

\item{...}{Other parameters passed to
\code{\link[bigmemory]{big.matrix}}. Note that
\code{nrow}, \code{ncol} and \code{type} of the
resulting \code{big.matrix} are automatically defined.}
}
\value{
The transposed big.matrix
}
\description{
This function implements a simple
cache-oblivious algorithm for the transposition of
a "big.matrix".
}
\examples{
X <- big.matrix(11, 23, shared = FALSE)
X[] <- rnorm(length(X))

Xt <- transpose(X, shared = FALSE)

print(all.equal(t(X[,]), Xt[,]))


}

