% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-assert.R
\name{without_downcast_warning}
\alias{without_downcast_warning}
\title{Temporarily disable downcast warning}
\usage{
without_downcast_warning(expr)
}
\arguments{
\item{expr}{The expression to evaluate without downcast warning.}
}
\value{
The result of the evaluated expression.
}
\description{
Temporarily disable downcast warning
}
\examples{

}
