% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach-descriptor.R
\name{big_attach}
\alias{big_attach}
\alias{big_attachExtdata}
\title{Attach a "big.matrix.descriptor"}
\usage{
big_attach(descriptorfile)

big_attachExtdata()
}
\arguments{
\item{descriptorfile}{Path to a ".desc" file.}
}
\value{
A \link[=big.matrix.descriptor-class]{big.matrix.descriptor} object.
}
\description{
Attach a "big.matrix.descriptor"
}
\examples{
# tmpFBM
tmpfile <- tempfile()
X.desc <- FBM(backingroot = basename(tmpfile),
              backingpath = dirname(tmpfile))(10, 10)

descriptorfile <- paste0(tmpfile, ".desc")
X.desc2 <- big_attach(descriptorfile)

all.equal(X.desc, X.desc2)
}
\keyword{internal}
