% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach-descriptor.R
\name{big_attach}
\alias{big_attach}
\alias{big_attachExtdata}
\title{Attach a "big.matrix.descriptor"}
\usage{
big_attach(descriptorfile)

big_attachExtdata()
}
\arguments{
\item{descriptorfile}{}
}
\value{
A \link[=big.matrix.descriptor-class]{big.matrix.descriptor} object.
}
\description{
Attach a "big.matrix.descriptor"
}
\examples{
tmpfile <- tempfile()
X.desc <- describe(
  big.matrix(10, 10, backingfile = basename(tmpfile),
             backingpath = dirname(tmpfile),
             descriptorfile = paste0(basename(tmpfile), ".desc"))
)
X.desc <- describe(test)
descriptorfile <- paste0(tmpfile, ".desc")
X.desc2 <- big_attach(descriptorfile)
all.equal(X.desc, X.desc2)
}
\keyword{internal}
