% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM-attach.R
\name{big_attach}
\alias{big_attach}
\alias{big_attachExtdata}
\title{Attach a Filebacked Big Matrix}
\usage{
big_attach(rdsfile)

big_attachExtdata()
}
\arguments{
\item{rdsfile}{Path to a ".rds" file.}
}
\value{
The \link{FBM} object stored in the rdsfile.
}
\description{
Attach a Filebacked Big Matrix
}
\examples{
# temporary FBM
X <- FBM(10, 10)$save()

rdsfile <- sub_bk(X$backingfile, ".rds")
X2 <- big_attach(rdsfile)

all.equal(X[], X2[])
}
\keyword{internal}
