% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSA.R
\name{big_CMSA}
\alias{big_CMSA}
\title{Cross-Model Selection and Averaging.}
\usage{
big_CMSA(FUN, X, y, ind.train = seq(nrow(X)), covar = NULL, K = 10,
  ncores = 1, method = c("geometric-median", "mean-wise", "median-wise"),
  block.size = 1000, ...)
}
\arguments{
\item{FUN}{Function that computes a linear scores for different
values along a regularization path. For now, this is relevant to
\link{big_spLinReg}, \link{big_spLogReg} and \link{big_spSVM}.
The corresponding \code{<FUN>.predict} function should exists.}

\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all rows are used.}

\item{K}{Number of folds that are used.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}

\item{method}{Method for combining vectors of coefficients. The default uses
the \href{https://en.wikipedia.org/wiki/Geometric_median}{geometric median}.}

\item{block.size}{Maximum number of columns read at once.
Default is \code{1000}. This parameter controls the trade-off between
memory usage and speed. Basically, the more you can load at once,
the quicker will be the execution time, at the expense of memory usage.}

\item{...}{Extra parameters to be passed to FUN.}

\item{feval}{Custimized evaluation function. You should always aim at
maximizing this function. If \code{feval} is like a loss function (which you
want to minimize), you should use its opposite instead.
Its only two arguments should be \code{pred} and \code{target}
(e.g. \code{function(pred, target) aucSample(pred, target, nsim = 1e5)}).}
}
\value{
A vector of resulting coefficients (intercept excluded), which
corresponds to the training set. You might want to recompute an optimal
intercept if needed.
}
\description{
\enumerate{
\item This function separates the training set in \code{K} folds.
\item \strong{In turn},
\itemize{
\item each fold is considered as an inner validation set and the others
(K - 1) folds form an inner training set,
\item the model is trained on the inner training set and the corresponding
predictions (scores) for the inner validation set are computed,
\item the vector of coefficients corresponding to the vector of scores which
maximizes \code{feval} is chosen.
}
\item The \code{K} resulting vectors of coefficients are then combined into one
vector (see the \code{method} parameter).
}
}
