% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colstats.R
\name{big_colstats}
\alias{big_colstats}
\title{Standard univariate statistics}
\usage{
big_colstats(X, ind.train = seq_len(nrow(X)))
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}
}
\value{
List of two numeric vectors \code{sum} and \code{var} with the
corresponding column statistics.
}
\description{
Standard \strong{univariate statistics} for columns of a big.matrix.
For now, the \code{sum} and \code{var} are implemented
(the \code{mean} and \code{sd} can easily be deduced, see examples).
}
\examples{
# Simulating some data
X <- big.matrix(41, 7)
X[] <- rnorm(length(X))

# Check the results
print(test <- big_colstats(X))

# Only with the first 10 rows
ind <- 1:10
print(test <- big_colstats(X, ind))
print(all.equal(test$sum, colSums(X[ind, ])))
print(all.equal(test$var, apply(X[ind, ], 2, var)))

# deduce mean and sd
# note that the are also implemented in big_scale()
means <- test$sum / length(ind) # if using all rows,
                                # divide by nrow(X) instead
print(all.equal(means, colMeans(X[ind, ])))
sds <- sqrt(test$var)
print(all.equal(sds, apply(X[ind, ], 2, sd)))

# NA in, NA out
X[2, 5] <- NA
print(big_colstats(X))
}
\seealso{
\link{colSums} \link{apply}
}
