% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM-copy-coerce.R
\docType{methods}
\name{big_copy}
\alias{big_copy}
\alias{big_copy,matrix-method}
\alias{big_copy,FBM-method}
\alias{big_copy,big.matrix-method}
\title{Methods to copy an object to a FBM}
\usage{
big_copy(x, ...)

\S4method{big_copy}{matrix}(x, ...)

\S4method{big_copy}{FBM}(x, type = names(x$type), ...)

\S4method{big_copy}{big.matrix}(x, ...)
}
\arguments{
\item{x}{Either a \link[=FBM-class]{FBM}, a \link[=big.matrix-class]{big.matrix} or
a standard R matrix.}
}
\description{
Methods to copy an object to a FBM
}
