% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{big_cor}
\alias{big_cor}
\title{Correlation}
\usage{
big_cor(X, ind.row = rows_along(X), ind.col = cols_along(X),
  block.size = 1000)
}
\arguments{
\item{X}{A \link{FBM}.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. \strong{Don't use negative indices.}}

\item{block.size}{Maximum number of columns read at once. Default is \code{1000}.}
}
\value{
The correlation matrix
}
\description{
Compute the correlation matrix of a Filebacked Big Matrix.
}
\details{
This algorithm is not really memory efficient. I'm planning on fixing this,
if it is really needed.
}
\examples{
X <- big_attachExtdata()

# Comparing with cor
K <- big_cor(X)
dim(K)

true <- cor(X[])
all.equal(K, true)

# Using only half of the data
n <- nrow(X)
ind <- sort(sample(n, n/2))
K2 <- big_cor(X, ind.row = ind)
dim(K2)

true2 <- cor(X[ind, ])
all.equal(K2, true2)
}
\seealso{
\link{cor}
}
