% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mult.R
\name{big_cprodMat}
\alias{big_cprodMat}
\title{Cross-product with a matrix}
\usage{
big_cprodMat(X., A, ind.row = rows_along(X.), ind.col = cols_along(X.),
  block.size = 1000, ncores2 = 1)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{A}{A matrix with \code{length(ind.row)} rows.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. Don't use negative indices.}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. Don't use negative indices.}

\item{block.size}{Maximum number of columns read at once.
Default is \code{1000}. This parameter controls the trade-off between
memory usage and speed. Basically, the more you can load at once,
the quicker will be the execution time, at the expense of memory usage.}

\item{ncores2}{Number of cores used. Default doesn't use parallelism.
For this function, use only half of the cores you have.}
}
\value{
\eqn{X^T /cdot A}.
}
\description{
Cross-product between a "big.matrix" and a matrix.
}
\examples{
X.desc <- big_attachExtdata()
n <- nrow(X.desc)
m <- ncol(X.desc)
A <- matrix(0, n, 10); A[] <- rnorm(length(A))

test <- big_cprodMat(X.desc, A)
true <- crossprod(attach.BM(X.desc)[,], A)
all.equal(test, true)

# subsetting
ind.row <- sample(n, n/2)
ind.col <- sample(m, m/2)

# test2 <- big_cprodMat(X.desc, A, ind.row, ind.col)
# returns an error. You need to use the subset of A:
test2 <- big_cprodMat(X.desc, A[ind.row, ], ind.row, ind.col)
true2 <- crossprod(attach.BM(X.desc)[ind.row, ind.col], A[ind.row, ])
all.equal(test2, true2)
}
