% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{big_increment}
\alias{big_increment}
\title{Increment an FBM}
\usage{
big_increment(BM, source)
}
\arguments{
\item{BM}{An \code{FBM} of type double.}

\item{source}{A matrix of same size as the \code{FBM} to increment.}
}
\value{
Returns nothing (\code{NULL}, invisibly).
}
\description{
Increment an FBM
}
\examples{
X <- FBM(10, 10, init = 0)
mat <- matrix(rnorm(100), 10, 10)

big_increment(X, mat)
all.equal(X[], mat)

big_increment(X, mat)
all.equal(X[], 2 * mat)

}
