% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mult.R
\name{big_prodMat}
\alias{big_prodMat}
\title{Product with a matrix}
\usage{
big_prodMat(X, A.col, ind.row = rows_along(X), ind.col = cols_along(X),
  block.size = 1000, ncores = 1)
}
\arguments{
\item{X}{A \link{FBM}.}

\item{A.col}{A matrix with \code{length(ind.col)} rows.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. \strong{Don't use negative indices.}}

\item{block.size}{Maximum number of columns read at once. Default is \code{1000}.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.
We advise you to use only half of the cores you have.}
}
\value{
\eqn{X \cdot A}.
}
\description{
Product between a Filebacked Big Matrix and a matrix.
}
\examples{
X.desc <- big_attachExtdata()
n <- nrow(X.desc)
m <- ncol(X.desc)
A <- matrix(0, m, 10); A[] <- rnorm(length(A))

test <- big_prodMat(X.desc, A)
true <- attach.BM(X.desc)[,] \%*\% A
all.equal(test, true)

# subsetting
ind.row <- sample(n, n/2)
ind.col <- sample(m, m/2)

\dontrun{test2 <- big_prodMat(X.desc, A, ind.row, ind.col)}
# returns an error. You need to use the subset of A:
test2 <- big_prodMat(X.desc, A[ind.col, ], ind.row, ind.col)
true2 <- attach.BM(X.desc)[ind.row, ind.col] \%*\% A[ind.col, ]
all.equal(test2, true2)
}
