% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mult.R
\name{big_prodVec}
\alias{big_prodVec}
\title{Product with a vector}
\usage{
big_prodVec(X., y, ind.row = rows_along(X.), ind.col = cols_along(X.))
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. Don't use negative indices.}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. Don't use negative indices.}
}
\value{
\eqn{X /cdot y}.
}
\description{
Product between a "big.matrix" and a vector.
}
\examples{
X.desc <- big_attachExtdata()
n <- nrow(X.desc)
m <- ncol(X.desc)
y <- rnorm(m)

test <- big_prodVec(X.desc, y)      # vector
true <- attach.BM(X.desc)[,] \%*\% y  # one-column matrix
all.equal(test, as.numeric(true))

# subsetting
ind.row <- sample(n, n/2)
ind.col <- sample(m, m/2)

# test2 <- big_prodVec(X.desc, y, ind.row, ind.col)
# returns an error. You need to use the subset of y:
test2 <- big_prodVec(X.desc, y[ind.col], ind.row, ind.col)
true2 <- attach.BM(X.desc)[ind.row, ind.col] \%*\% y[ind.col]
all.equal(test2, as.numeric(true2))
}
