% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomSVD.R
\name{big_randomSVD}
\alias{big_randomSVD}
\title{Randomized SVD}
\usage{
big_randomSVD(X., fun.scaling, ind.row = rows_along(X.),
  ind.col = cols_along(X.), k = 10, tol = 1e-04, verbose = FALSE,
  ncores = 1)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{fun.scaling}{A function that returns a named list of
\code{mean} and \code{sd} for every column, to scale each of their elements
such as followed: \deqn{\frac{X_{i,j} - mean_j}{sd_j}}.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. \strong{Don't use negative indices.}}

\item{k}{Number of singular vectors/values to compute. Default is \code{10}.
\strong{This algorithm should be used to compute only a
few singular vectors/values.}}

\item{tol}{Precision parameter of \link[RSpectra:svds]{svds}.
Default is \code{1e-4}.}

\item{verbose}{Should some progress be printed? Default is \code{FALSE}.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}
}
\value{
A named list (an S3 class "big_SVD") of
\itemize{
\item \code{d}, the singular values,
\item \code{u}, the left singular vectors,
\item \code{v}, the right singular vectors,
\item \code{niter}, the number of the iteration of the algorithm,
\item \code{nops}, number of Matrix-Vector multiplications used,
\item \code{means}, the centering vector,
\item \code{sds}, the scaling vector.
}

Note that to obtain the Principal Components, you must use
\link[=predict.big_SVD]{predict} on the result. See examples.
}
\description{
An algorithm for SVD (or PCA) of a \code{big.matrix} based on the algorithm
in RSpectra (by Yixuan Qiu and Jiali Mei).
\cr
This algorithm is linear in time in all dimensions and is very
memory-efficient. Thus, it can be used on very large big.matrices.
}
\note{
The idea of using this Implicitly Restarted Arnoldi Method algorithm
comes from G. Abraham, Y. Qiu, and M. Inouye,
FlashPCA2: principal component analysis of biobank-scale genotype datasets,
bioRxiv: \url{https://doi.org/10.1101/094714}.
\cr
It proved to be faster than our implementation of the "blanczos" algorithm
in Rokhlin, V., Szlam, A., & Tygert, M. (2010).
A Randomized Algorithm for Principal Component Analysis.
SIAM Journal on Matrix Analysis and Applications, 31(3), 1100–1124.
\url{http://dx.doi.org/10.1137/080736417}.
}
\examples{
X.desc <- big_attachExtdata()
K <- 10

# Using only half of the data for "training"
n <- nrow(X.desc)
ind <- sort(sample(n, n/2))
test <- big_randomSVD(X.desc, fun.scaling = big_scale(), ind.row = ind, k = K)
str(test)
plot(test$u)

pca <- prcomp(attach.BM(X.desc)[ind, ], center = TRUE, scale. = TRUE)

# same scaling
all.equal(test$means, pca$center)
all.equal(test$sds, pca$scale)

# use this function to predict scores
class(test)
scores <- predict(test)
# scores and loadings are the same or opposite
plot(scores, pca$x[, 1:K])
plot(test$v, pca$rotation[, 1:K])

# projecting on new data
ind2 <- setdiff(rows_along(X.desc), ind)
scores.test2 <- predict(test, X.desc, ind.row = ind2)
scores.test3 <- predict(pca, attach.BM(X.desc)[-ind, ])
plot(scores.test2, scores.test3[, 1:K])

}
\seealso{
\link[RSpectra:svds]{svds}
}
