% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{big_read}
\alias{big_read}
\title{Read a file}
\usage{
big_read(file, sep, header, nlines, confirmed = FALSE, verbose = TRUE,
  ind.skip = integer(0), ind.meta = integer(0), nlines.block = NULL,
  fun.con = function(f) file(f, open = "rt"), type = NULL, ...)
}
\arguments{
\item{file}{The path to the file to be read.}

\item{sep}{The field separator character.}

\item{header}{A logical value indicating whether the file contains the names
of the variables as its first line.}

\item{nlines}{Number of lines of the file (including possible header).
You can use \code{bigstatsr::nlines} or \code{fpeek::peek_count_lines} to get it.}

\item{confirmed}{A logical value indicating whether you don't want to pause
to confirm that everything has been parsed like you wanted.
Default is \code{FALSE} and will ask you to confirm during execution.}

\item{verbose}{Be verbose? Default is \code{TRUE}.}

\item{ind.skip}{An integer vector specifying which columns to skip.}

\item{ind.meta}{An integer vector specifying which columns should be part of
the meta information. Non-numeric columns will automatically be added to
meta information because they can't be stored in an FBM.}

\item{nlines.block}{Number of lines to be read at once. Default automatically
gets this number based on \code{getOption("bigstatsr.block.sizeGB")}.}

\item{fun.con}{Function to open a connection from a file.}

\item{type}{Type of the resulting Filebacked Big Matrix.
Default is automatically guessing between \code{"integer"} and \code{"double"}.}

\item{...}{Arguments passed on to \code{FBM}
\describe{
  \item{backingfile}{Path to the file storing the Big Matrix on disk.
An extension ".bk" will be automatically added. Default stores in the
temporary directory.}
  \item{create_bk}{Create a backingfile (the default) or use an existing one
(which should be named by the \code{backingfile} parameter and have an
extension ".bk"). For example, this could be used to convert a filebacked
\code{big.matrix} from package \strong{bigmemory} to a \link[=FBM-class]{FBM}.}
  \item{save}{Whether to save the result object in an ".rds" file alongside
the backingfile. Default is \code{FALSE}.}
}}
}
\value{
A list with 3 elements:
\itemize{
\item A Filebacked Big Matrix.
\item A vector of colnames.
\item A data frame of meta information (e.g. non-numeric variables such as rownames)
}
}
\description{
Read a file as a Filebacked Big Matrix and a data frame of meta information.
For a mini-tutorial, please see \href{https://goo.gl/91oNxU}{this vignette}.
}
