% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBM.R
\name{big_readBM}
\alias{big_readBM}
\title{Read a file}
\usage{
big_readBM(file, file.nheader = 0, file.nline = NULL, info.nelem = 0,
  split = " ", read.what = double(), read.transfo = identity,
  BM.type = typeof(read.what), transpose = FALSE, ...)
}
\arguments{
\item{file}{The path to the file to be read.}

\item{file.nheader}{Number of lines to read at the beginning of the file,
as a separate header information. Default is \code{0}. Each line is read as one
string. You may need to postprocess it with \link{strsplit}. See examples.}

\item{file.nline}{Number of total lines of the file. Default is \code{NULL} and a
function computes it. This function doesn't work for compressed files so that
you will have to explicitly specify the number of lines of the file.}

\item{info.nelem}{Number of elements of extra information to read at the
beginning of each line. Default is \code{0}.}

\item{split}{The separator used in the file. Default is a space.}

\item{read.what}{What type of elements to scan? Default is \code{double()}.
You can also use \code{character()} or \code{integer()}.}

\item{read.transfo}{Function that transforms each line you read.}

\item{BM.type}{Type of the resulting Filebacked Big Matrix. Default uses the
type of \code{read.what}. This can be useful if you have only small integers or
that the \code{read.transfo} function transforms the type of what is read.}

\item{transpose}{Should the resulting Filebacked Big Matrix be transposed?
Default is \code{FALSE}.}
}
\value{
A Filebacked Big Matrix with two attributes \code{header} and \code{info}.
}
\description{
Read a file as a Filebacked Big Matrix. For a mini-tutorial, please see the
\href{https://privefl.github.io/bigstatsr/articles/read-BM-from-file.html}{vignette}.
}
