% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{big_scale}
\alias{big_scale}
\title{Some scaling functions}
\usage{
big_scale(center = TRUE, scale = TRUE)
}
\arguments{
\item{center}{A logical value: whether to return means or 0s.}

\item{scale}{A logical value: whether to return sds or 1s. \strong{You can't
use scale without using center.}}
}
\value{
A new \strong{function} that returns a named list of two vectors "mean" and "sd"
which are as long as the number of columns of \code{X}.
}
\description{
Some scaling functions for a "big.matrix" to be used as
the \code{fun.scaling} parameter of some functions of this package.
}
\examples{
# Simulating some data
X <- big.matrix(41, 17)
X[] <- rnorm(length(X))

# No scaling
big_noscale <- big_scale(center = FALSE, scale = FALSE)
print(class(big_noscale)) # big_scale returns a new function
print(big_noscale(X))
big_noscale2 <- big_scale(center = FALSE)
print(big_noscale2(X)) # you can't scale without centering

# Centering
big_center <- big_scale(scale = FALSE)
print(big_center(X))
# + scaling
print(test <- big_scale()(X, 1:5)) # note that big_scale() is a function
print(all.equal(test$mean, colMeans(X[1:5, ])))
print(all.equal(test$sd, apply(X[1:5, ], 2, sd)))
}
\seealso{
\link{scale}
}

