% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglasso.R
\name{big_spRegLogCV2}
\alias{big_spRegLogCV2}
\title{Title}
\usage{
big_spRegLogCV2(X, y01.train, ind.train = seq(nrow(X)), covar.train = NULL,
  K = 10, ncores = 1, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
