% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD.R
\name{big_SVD}
\alias{big_SVD}
\title{SVD}
\usage{
big_SVD(X., fun.scaling, ind.row = rows_along(X.), block.size = 1000,
  k = NULL, thr.eigval = 1e-04)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{fun.scaling}{A function that returns a named list of
\code{mean} and \code{sd} for every column, to scale each of their elements
such as followed: \deqn{\frac{X_{i,j} - mean_j}{sd_j}}.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{block.size}{Maximum number of columns read at once. Default is \code{1000}.}

\item{k}{Number of singular vectors/values to compute. Default is all.}

\item{thr.eigval}{Threshold to remove "unsignificant" singular vectors.
Default is \code{1e-4}.}
}
\value{
A named list (an S3 class "big_SVD") of
\itemize{
\item \code{d}, the singular values,
\item \code{u}, the left singular vectors,
\item \code{v}, the right singular vectors,
\item \code{means}, the centering vector,
\item \code{sds}, the scaling vector.
}

Note that to obtain the Principal Components, you must use
\link[=predict.big_SVD]{predict} on the result. See examples.
}
\description{
An algorithm for SVD (or PCA) of a \code{big.matrix} through the eigen
decomposition of the covariance between variables (primal)
or observations (dual).
}
\details{
To get \eqn{X = U \cdot D \cdot V^T},
\itemize{
\item if the number of observation is small, this function computes
\eqn{K_(2) = X \cdot X^T = U \cdot D^2 \cdot U^T} and then
\eqn{V = X^T \cdot U \cdot D^{-1}},
\item if the number of variable is small, this function computes
\eqn{K_(1) = X^T \cdot X = V \cdot D^2 \cdot V^T} and then
\eqn{U = X \cdot V \cdot D^{-1}},
\item if both dimensions are large and you only want a partial SVD,
please use \link{big_randomSVD} instead.
}
}
\examples{
X.desc <- big_attachExtdata()
n <- nrow(X.desc)

# Using only half of the data
ind <- sort(sample(n, n/2))

test <- big_SVD(X. = X.desc, fun.scaling = big_scale(), ind.row = ind)
str(test)
plot(test$u)

pca <- prcomp(attach.BM(X.desc)[ind, ], center = TRUE, scale. = TRUE)

# same scaling
all.equal(test$means, pca$center)
all.equal(test$sds, pca$scale)

# scores and loadings are the same or opposite
# except for last eigenvalue which is equal to 0
# due to centering of columns
scores <- test$u \%*\% diag(test$d)
class(test)
scores2 <- predict(test) # use this function to predict scores
all.equal(scores, scores2)
dim(scores)
dim(pca$x)
tail(pca$sdev)
plot(scores2, pca$x[, 1:ncol(scores2)])
plot(test$v[1:100, ], pca$rotation[1:100, 1:ncol(scores2)])

# projecting on new data
X2 <- sweep(sweep(attach.BM(X.desc)[-ind, ], 2, test$means, '-'),
            2, test$sds, '/')
scores.test <- X2 \%*\% test$v
ind2 <- setdiff(rows_along(X.desc), ind)
scores.test2 <- predict(test, X.desc, ind.row = ind2) # use this
all.equal(scores.test, scores.test2)
scores.test3 <- predict(pca, attach.BM(X.desc)[-ind, ])
plot(scores.test2, scores.test3[, 1:ncol(scores.test2)])
}
\seealso{
\link[stats:prcomp]{prcomp}
}
