% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrossprodSelf.R
\name{big_tcrossprodSelf}
\alias{big_tcrossprodSelf}
\title{Tcrossprod}
\usage{
big_tcrossprodSelf(X., fun.scaling, ind.row = rows_along(X.),
  ind.col = cols_along(X.), block.size = 1000)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{fun.scaling}{A function that returns a named list of
\code{mean} and \code{sd} for every column, to scale each of their elements
such as followed: \deqn{\frac{X_{i,j} - mean_j}{sd_j}.}}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. \strong{Don't use negative indices.}}

\item{block.size}{Maximum number of columns read at once. Default is \code{1000}.}
}
\value{
A matrix, \eqn{X.row X.row^T}, with the following two attributes:
\itemize{
\item a numeric vector \code{mean} of column scaling,
\item a numeric vector \code{sd} of column scaling.
}
}
\description{
Compute \eqn{X.row X.row^T} for a \code{big.matrix} \code{X}
after applying a particular scaling to it.
}
\details{
This algorithm is not really memory efficient. I'm planning on fixing this,
if it is really needed.
}
\examples{
X.desc <- big_attachExtdata()

# Comparing with tcrossprod
big_noscale <- big_scale(center = FALSE)
K <- big_tcrossprodSelf(X.desc, fun.scaling = big_noscale)
dim(K)

true <- tcrossprod(attach.BM(X.desc)[,])
all.equal(K, true, check.attributes = FALSE)

# Using only half of the data
n <- nrow(X.desc)
ind <- sort(sample(n, n/2))
K2 <- big_tcrossprodSelf(X.desc, fun.scaling = big_noscale,
                            ind.row = ind)
dim(K2)

true2 <- tcrossprod(attach.BM(X.desc)[ind, ])
all.equal(K2, true2, check.attributes = FALSE)
}
\seealso{
\link{tcrossprod}
}
