% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{big_transpose}
\alias{big_transpose}
\title{Transposition}
\usage{
big_transpose(X., fun.createBM = BM())
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{fun.createBM}{A function that creates a \code{big.matrix}. It should have
exactly 3 parameters \code{nrow}, \code{ncol} and \code{type}. See \link{FBM}.}
}
\value{
The new transposed \code{big.matrix} (or its descriptor). Its dimensions
and type are automatically determined from the input \code{big.matrix}.
}
\description{
This function implements a simple cache-oblivious algorithm for
the transposition of a "big.matrix".
}
\examples{
X.desc <- big_attachExtdata()
Xt.desc <- big_transpose(X.desc, fun.createBM = tmpFBM())
identical(t(attach.BM(X.desc)[,]), attach.BM(Xt.desc)[,])

}
