% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{big_transpose}
\alias{big_transpose}
\title{Transposition}
\usage{
big_transpose(X, ...)
}
\arguments{
\item{X}{A \link{FBM}.}
}
\value{
The new transposed Filebacked Big Matrix (or its descriptor). Its dimensions
and type are automatically determined from the input Filebacked Big Matrix.
}
\description{
This function implements a simple cache-oblivious algorithm for
the transposition of a Filebacked Big Matrix.
}
\examples{
X <- big_copy(matrix(rnorm(150), 10, 15))
Xt <- big_transpose(X)
identical(t(X[]), Xt[])

X <- big_attachExtdata()
Xt <- big_transpose(X)
identical(t(X[]), Xt[])

}
