% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{big_transpose}
\alias{big_transpose}
\title{Transposition of a "big.matrix".}
\usage{
big_transpose(X, ...)
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{...}{Arguments passed on to \code{bigmemory::big.matrix}
\describe{
  \item{dimnames}{a list of the row and column names; use with caution 
for large objects.}
  \item{separated}{use separated column organization of the data; 
see details.}
  \item{backingfile}{the root name for the file(s) for the cache of \code{x}.}
  \item{backingpath}{the path to the directory containing the file 
backing cache.}
  \item{descriptorfile}{the name of the file to hold the backingfile 
description, for subsequent use with \code{\link{attach.big.matrix}}; 
if \code{NULL}, the \code{backingfile} is used as the root part of the 
descriptor file name.  The descriptor file is placed in the same directory 
as the backing files.}
  \item{binarydescriptor}{the flag to specify if the binary RDS format 
should be used for the backingfile description, for subsequent use with 
\code{\link{attach.big.matrix}}; if \code{NULL} of \code{FALSE}, the 
\code{dput()} file format is used.}
  \item{shared}{\code{TRUE} by default, and always \code{TRUE} if the 
\code{big.matrix} is file-backed.  For a non-filebacked \code{big.matrix}, 
\code{shared=FALSE} uses non-shared memory, which can be more stable for 
large (say, >50\% of RAM) objects.  Shared memory allocation can sometimes 
fail in such cases due to exhausted shared-memory resources in the system.}
}}
}
\value{
The transposed big.matrix. Its dimension and type
are automatically determined from the input \code{big.matrix}.
}
\description{
This function implements a simple cache-oblivious
algorithm for the transposition of a "big.matrix".
}
\examples{
X <- big.matrix(11, 23, shared = FALSE)
X[] <- rnorm(length(X))

Xt <- big_transpose(X, shared = FALSE)

print(identical(
  t(X[,]),
  Xt[,]
))
}

