% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univLogReg.R
\name{big_univLogReg}
\alias{big_univLogReg}
\title{Column-wise logistic regression}
\usage{
big_univLogReg(X., y01.train, ind.train = rows_along(X.),
  ind.col = cols_along(X.), covar.train = NULL, ncores2 = 1,
  tol = 1e-08, maxiter = 20)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{y01.train}{Vector of responses, corresponding to \code{ind.train}.
Must be 0s and 1s.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all rows are used.
\strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. \strong{Don't use negative indices.}}

\item{covar.train}{Matrix of covariables to be added in each model to correct
for confounders (e.g. the scores of PCA), corresponding to \code{ind.train}.
Default is \code{NULL} and corresponds to only adding an intercept to each model.}

\item{ncores2}{Number of cores used. Default doesn't use parallelism.
For this function, use only half of the cores you have.}

\item{tol}{Relative tolerance to assess convergence of the coefficient.
Default is \code{1e-8}.}

\item{maxiter}{Maximum number of iterations before giving up.
Default is \code{20}. Usually, convergence is reached within 3 or 4 iterations.
If there is not convergence,
\link[stats:glm]{glm} is used instead for the corresponding column.}
}
\value{
A data.frame with as elements:
\enumerate{
\item the slopes of each regression,
\item the standard errors of each slope,
\item the number of iteration for each slope. If is \code{NA}, this means that the
algorithm didn't converge, and \link[stats:glm]{glm} was used instead.
\item the z-scores associated with each slope,
\item the p-values associated with each z-score.
}
}
\description{
Slopes of column-wise logistic regressions of each column
of a \code{big.matrix}, with some other associated statistics.
Covariates can be added to correct for confounders.
}
\examples{
X.desc <- big_attachExtdata()
n <- nrow(X.desc)
y01 <- sample(0:1, size = n, replace = TRUE)
covar <- matrix(rnorm(n * 3), n)

X1 <- attach.BM(X.desc)[, 1] # only first column of the `big.matrix`

# Without covar
str(big_univLogReg(X.desc, y01))
summary(glm(y01 ~ X1, family = "binomial"))$coefficients[2, ]

# With all data
str(big_univLogReg(X.desc, y01, covar.train = covar))
summary(glm(y01 ~ X1 + covar, family = "binomial"))$coefficients[2, ]

# With only half of the data
ind.train <- sort(sample(n, n/2))
str(big_univLogReg(X.desc, y01[ind.train],
                   covar.train = covar[ind.train, ],
                   ind.train = ind.train))
summary(glm(y01 ~ X1 + covar, family = "binomial",
            subset = ind.train))$coefficients[2, ]
}
\seealso{
\link[stats:glm]{glm}
}
