% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigstatsr.R
\name{bigstatsr-package}
\alias{bigstatsr}
\alias{bigstatsr-package}
\title{Statistical Tools for bigmemory Matrices.}
\arguments{
\item{x}{A \code{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part.
If not specified, all data are used.}
}
\description{
Easy-to-use, efficient, flexible and scalable
statistical tools. Use bigmemory matrices.
}
\keyword{package}

