% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigstatsr.R
\docType{package}
\name{bigstatsr-package}
\alias{bigstatsr}
\alias{bigstatsr-package}
\title{bigstatsr: Statistical Tools for Bigmemory Matrices}
\arguments{
\item{X}{A \link[=big.matrix-class]{big.matrix}.}

\item{X.desc}{A \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}

\item{X.code}{Either a \link[=BM.code-class]{BM.code} or
a \link[=BM.code.descriptor-class]{BM.code.descriptor}.}

\item{y.train}{Vector of responses, corresponding to \code{ind.train}.}

\item{y01.train}{Vector of responses, corresponding to \code{ind.train}.
\strong{Must be only 0s and 1s.}}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all rows are used.
\strong{Don't use negative indices.}}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. \strong{Don't use negative indices.}}

\item{block.size}{Maximum number of columns read at once. Default is \code{1000}.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.
We advise you to use only half of the cores you have.}

\item{fun.scaling}{A function that returns a named list of
\code{mean} and \code{sd} for every column, to scale each of their elements
such as followed: \deqn{\frac{X_{i,j} - mean_j}{sd_j}.}}

\item{fun.createBM}{A function that creates a \code{big.matrix}. It should have
exactly 3 parameters \code{nrow}, \code{ncol} and \code{type}. See \link{FBM}.}

\item{covar.train}{Matrix of covariables to be added in each model to correct
for confounders (e.g. the scores of PCA), corresponding to \code{ind.train}.
Default is \code{NULL} and corresponds to only adding an intercept to each model.}

\item{covar.row}{Matrix of covariables to be added in each model to correct
for confounders (e.g. the scores of PCA), corresponding to \code{ind.row}.
Default is \code{NULL} and corresponds to only adding an intercept to each model.}

\item{thr.eigval}{Threshold to remove "unsignificant" singular vectors.
Default is \code{1e-4}.}
}
\description{
Easy-to-use, efficient, flexible and scalable
statistical tools. Use bigmemory matrices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://privefl.github.io/bigstatsr}
  \item Report bugs at \url{https://github.com/privefl/bigstatsr/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

Other contributors:
\itemize{
  \item Michael Blum \email{michael.blum@univ-grenoble-alpes.fr} [thesis advisor]
  \item Hugues Aschard \email{hugues.aschard@pasteur.fr} [thesis advisor]
}

}
