% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigXtX.R
\name{BigXtX}
\alias{BigXtX}
\title{Crossprod for a big.matrix.}
\usage{
BigXtX(X, block.size, ind.train = seq(nrow(X)), vec.center = rep(0,
  ncol(X)), vec.scale = rep(1, ncol(X)), progress = TRUE)
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{block.size}{Maximum number of columns read at once.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part.
If not specified, all data are used.}

\item{vec.center}{Vector that will be subtracted to the matrix,
columnwise. Typically, the mean of each column.
See \code{\link{colmeans}}.}

\item{vec.scale}{Vector that will be divided to the matrix
(after the substraction), columnwise.
Typically, the sd of each column. See \code{\link{colsds}}.}

\item{progress}{Use a progress bar for the computation
of the correlation matrix? Default is \code{TRUE}.}
}
\value{
\eqn{X.train^T X.train}
}
\description{
Compute \eqn{X^T X} for a "big.matrix" X
after applying a particular scaling to it.
}
\examples{
# Simulating some data
X <- big.matrix(17, 41, shared = FALSE)
X[] <- rnorm(length(X))


# center and scale
vec.center <- colmeans(X)
vec.scale <- colsds(X)

# Comparing with tcrossprod
test <- BigXtX(X = X,
               block.size = 10,
               vec.center = vec.center,
               vec.scale = vec.scale)

mat <- sweep(sweep(X[,], 2, vec.center, '-'), 2, vec.scale, '/')
diff <- test[,] - crossprod(mat)
print(max(abs(diff)))


# Using only half of the data for "training"
ind <- sort(sample(nrow(X), nrow(X)/2))
vec.center <- colmeans(X, ind)
vec.scale <- colsds(X, ind)

test <- BigXtX(X = X,
               block.size = 10,
               ind.train = ind,
               vec.center = vec.center,
               vec.scale = vec.scale)

mat <- sweep(sweep(X[ind, ],  2, vec.center, '-'), 2, vec.scale, '/')
diff <- test[,] - crossprod(mat)
print(max(abs(diff)))
}
\seealso{
\code{\link{crossprod}}
}

