% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM-code256.R
\docType{class}
\name{FBM.code256-class}
\alias{FBM.code256-class}
\alias{FBM.code256_RC}
\alias{FBM.code256}
\alias{[,FBM.code256,ANY,ANY,ANY-method}
\alias{add_code256}
\title{Class FBM.code256}
\usage{
FBM.code256(nrow, ncol, code, ...)

\S4method{[}{FBM.code256,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

add_code256(x, code, save = FALSE)
}
\arguments{
\item{code}{A numeric vector (of length 256).
You should contruct it with \code{rep(NA_real_, 256)} and then replace the values
which are of interest for you.}

\item{x}{A \link[=FBM.code256-class]{FBM.code256}.}
}
\description{
A reference class for storing and accessing up to 256 arbitrary different
values using a Filebacked Big Matrix of type \code{unsigned char}. Compared to a
\link[=FBM-class]{Filebacked Big Matrix}, it adds a slot \code{code} which is used as
a lookup table of size 256.

Wrapper constructor for class \code{FBM.code256}.

Accessor method for class \code{FBM.code256}.

Converter from class \code{FBM} to \code{FBM.code256}.
}

\examples{
X <- FBM(10, 10, type = "raw")
X[] <- sample(as.raw(0:3), size = length(X), replace = TRUE)
X[]

code <- rep(NA_real_, 256)
code[1:3] <- c(1, 3, 5)

X.code <- add_code256(X, code)
X.code[]

# Or directly
X.code2 <- FBM.code256(10, 10, code, init = sample(as.raw(0:3), 100, TRUE))
X.code2[]

}
