% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelRandomProjPCA.R
\name{ParallelRandomProjPCA}
\alias{ParallelRandomProjPCA}
\title{Title}
\usage{
ParallelRandomProjPCA(X, fun.scaling, block.size = 1000, K = 10, I = 10,
  use.Eigen = TRUE, backingpath = NULL, tmp.lock.name = "file.lock",
  TIME = 0.01, ncores)
}
\arguments{
\item{ncores}{}
}
\description{
Title
}

