% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mhtest}
\alias{plot.mhtest}
\title{Plot method}
\usage{
\method{plot}{mhtest}(x, type = c("Manhattan", "Volcano"),
  main = paste(type, "Plot"), coeff = 1, ...)
}
\arguments{
\item{x}{An object of class \code{mhtest}.}

\item{type}{Either
\itemize{
\item "Manhattan": plot of the negative logarithm (in base 10) of p-values
(the default).
\item "Volcaco": plot of the negative logarithm of p-values against the
estimation of coefficients (e.g. betas in linear regression).
}}

\item{main}{The title of the plot. Default use the \code{type}.}

\item{coeff}{Relative size of text. Default is \code{1}.}

\item{...}{Not used.}
}
\value{
A \code{ggplot2} object. You can plot it using the \code{print} method.
You can modify it as you wish by adding layers. You might want to read
\href{http://r4ds.had.co.nz/data-visualisation.html}{this chapter}
to get more familiar with the package \strong{ggplot2}. \cr
Do you want dynamic plots? Go check \url{https://plot.ly/ggplot2}.
}
\description{
Plot method for class \code{mhtest}.
}
\examples{
X.desc <- big_attachExtdata()
n <- nrow(X.desc)
y <- rnorm(n)
test <- big_univLinReg(X.desc, y)

plot(test)
plot(test, type = "Volcano")

}
\seealso{
\link{big_univLinReg}, \link{big_univLogReg},
\link{plot.big_SVD} and \link{asPlotlyText}.
}
