% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhtest.R
\name{predict.mhtest}
\alias{predict.mhtest}
\title{Predict method}
\usage{
\method{predict}{mhtest}(object, scores = object$score, ...)
}
\arguments{
\item{object}{An object of class \code{mhtest} from you get the probability
function with possibily pre-transformation of scores.}

\item{scores}{Raw scores (before transformation) that you want to transform
to p-values.}

\item{...}{Not used.}
}
\value{
Vector of p-values associated with \code{scores} and \code{object}.
}
\description{
Predict method for class \code{mhtest}.
}
\seealso{
\link{big_univLinReg} and \link{big_univLogReg}.
}
