% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-parallelize.R, R/plot.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{nb_cores}
\alias{plot_grid}
\alias{rows_along}
\alias{cols_along}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bigparallelr}{\code{\link[bigparallelr]{nb_cores}}, \code{\link[bigparallelr]{rows_along}}, \code{\link[bigparallelr]{cols_along}}}

  \item{cowplot}{\code{\link[cowplot]{plot_grid}}}
}}

