% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM.R
\name{sub_bk}
\alias{sub_bk}
\title{Replace extension 'bk'}
\usage{
sub_bk(path, replacement = "")
}
\arguments{
\item{path}{String with extension 'bk'.}

\item{replacement}{Replacement of '.bk'. Default replaces by nothing.}
}
\value{
String with extension '.bk' replaced by \code{replacement}.
}
\description{
Replace extension 'bk'
}
\examples{
path <- "toto.bk"
sub_bk(path)
sub_bk(path, ".rds")
}
