% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.big_sp_list}
\alias{summary.big_sp_list}
\title{Summary method}
\usage{
\method{summary}{big_sp_list}(object, best.only = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{big_sp_list}.}

\item{best.only}{Whether to return only one row corresponding to the best
model? The best model is the one smallest \code{$validation_loss}.}

\item{...}{Not used.}
}
\value{
A tibble with, for each \code{$alpha}, a mean \code{$validation_loss}, a mean
vector of coefficients \code{$beta}, the corresponding number of non-zero
coefficients \code{$nb_var}, and the reasons of method completion \code{$message}.
}
\description{
Summary method for class \code{big_sp_list}.
}
