% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBM.R
\name{tmpFBM.rm}
\alias{tmpFBM.rm}
\title{Remove temporary backing files}
\usage{
tmpFBM.rm(X.)
}
\arguments{
\item{X.}{Either a \link[=big.matrix-class]{big.matrix} or
a \link[=big.matrix.descriptor-class]{big.matrix.descriptor}.}
}
\value{
\code{0} for success, \code{1} for failure, invisibly.
  Not deleting a non-existent file is not a failure, nor is being unable
  to delete a directory if \code{recursive = FALSE}.  However, missing
  values in \code{x} are regarded as failures.
}
\description{
Remove temporary backing files
}
\examples{
list.files(path = tempdir())
X.desc <- tmpFBM()(10, 10)
list.files(path = tempdir())

tmpFBM.rm(X.desc)
list.files(path = tempdir())

}
